Add files via upload
The attatched file code is applied on all the categorical features in our provided database and the target variable consists of the last column i.e. 'NObeyesdad' and the used algorithm is Supervised Machine Learning Algorithm (SVM)  that classifies data by finding an optimal line or hyperplane that maximizes the distance between each class in an N-dimensional space. Features responsible for the resultant accuracy of 88.89%  is regularity(C) set to 100, scale='gamma', and scaling features.